feat(HSU): move sig verification from auth to HSU
