+++ packages/auth/src/index.ts
@@ -18,7 +18,6 @@ import { createUnauthenticatedClient as createOpenPaymentsClient } from 'open-pa
 
 export { KeyInfo } from './accessToken/service'
 export { JWKWithRequired } from './client/service'
-export { HttpSigContext, verifySigAndChallenge } from './signature/middleware'
 const container = initIocContainer(Config)
 const app = new App(container)
