+++ packages/auth/src/signature/middleware.ts
@@ -1,64 +1,13 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
-import * as crypto from 'crypto'
-import { importJWK } from 'jose'
-import { JWK } from 'open-payments'
-import { verifyContentDigest } from 'httpbis-digest-headers'
+import {
+  verifySigFromClient,
+  validateHttpSigHeaders,
+  HttpSigContext
+} from 'http-signature-utils'
 
 import { AppContext } from '../app'
 import { Grant } from '../grant/model'
-import { Context } from 'koa'
-
-export async function verifySig(
-  sig: string,
-  jwk: JWK,
-  challenge: string
-): Promise<boolean> {
-  const publicKey = (await importJWK(jwk)) as crypto.KeyLike
-  const data = Buffer.from(challenge)
-  return crypto.verify(null, data, publicKey, Buffer.from(sig, 'base64'))
-}
-
-export async function verifySigAndChallenge(
-  clientKey: JWK,
-  ctx: HttpSigContext
-): Promise<boolean> {
-  const sig = ctx.headers['signature'] as string
-  const sigInput = ctx.headers['signature-input'] as string
-  const challenge = sigInputToChallenge(sigInput, ctx)
-  if (!challenge) {
-    ctx.throw(400, 'invalid signature input', { error: 'invalid_request' })
-  }
-
-  const verified = await verifySig(
-    sig.replace('sig1=', ''),
-    clientKey,
-    challenge
-  )
-
-  if (verified) {
-    return true
-  } else {
-    ctx.throw(401, 'invalid signature')
-  }
-}
-
-async function verifySigFromClient(
-  client: string,
-  ctx: HttpSigContext
-): Promise<boolean> {
-  const clientService = await ctx.container.use('clientService')
-  const clientKey = await clientService.getKey({
-    client,
-    keyId: ctx.clientKeyId
-  })
-
-  if (!clientKey) {
-    ctx.throw(400, 'invalid client', { error: 'invalid_client' })
-  }
-
-  return verifySigAndChallenge(clientKey, ctx)
-}
 
 async function verifySigFromBoundKey(
   grant: Grant,
@@ -73,18 +22,6 @@ async function verifySigFromBoundKey(
   return verifySigFromClient(grant.client, ctx)
 }
 
-// TODO: Replace with public httpsig library
-function getSigInputComponents(sigInput: string): string[] | null {
-  // https://datatracker.ietf.org/doc/html/rfc8941#section-4.1.1.1
-  const messageComponents = sigInput
-    .split('sig1=')[1]
-    ?.split(';')[0]
-    ?.split(' ')
-  return messageComponents
-    ? messageComponents.map((component) => component.replace(/[()"]/g, ''))
-    : null
-}
-
 const KEY_ID_PREFIX = 'keyid="'
 
 function getSigInputKeyId(sigInput: string): string | undefined {
@@ -93,94 +30,6 @@ function getSigInputKeyId(sigInput: string): string | undefined {
     .find((param) => param.startsWith(KEY_ID_PREFIX))
   // Trim prefix and quotes
   return keyIdParam?.slice(KEY_ID_PREFIX.length, -1)
-}
-
-function validateSigInputComponents(
-  sigInputComponents: string[],
-  ctx: Context
-): boolean {
-  // https://datatracker.ietf.org/doc/html/draft-ietf-gnap-core-protocol#section-7.3.1
-
-  for (const component of sigInputComponents) {
-    // https://datatracker.ietf.org/doc/html/draft-ietf-httpbis-message-signatures-09#section-2.1
-    if (component !== component.toLowerCase()) return false
-  }
-
-  const isValidContentDigest =
-    !sigInputComponents.includes('content-digest') ||
-    (!!ctx.headers['content-digest'] &&
-      ctx.request.body &&
-      Object.keys(ctx.request.body).length > 0 &&
-      sigInputComponents.includes('content-digest') &&
-      verifyContentDigest(
-        JSON.stringify(ctx.request.body),
-        ctx.headers['content-digest'] as string
-      ))
-
-  return !(
-    !isValidContentDigest ||
-    !sigInputComponents.includes('@method') ||
-    !sigInputComponents.includes('@target-uri') ||
-    (ctx.headers['authorization'] &&
-      !sigInputComponents.includes('authorization'))
-  )
-}
-
-export function sigInputToChallenge(
-  sigInput: string,
-  ctx: Context
-): string | null {
-  const sigInputComponents = getSigInputComponents(sigInput)
-
-  if (
-    !sigInputComponents ||
-    !validateSigInputComponents(sigInputComponents, ctx)
-  )
-    return null
-
-  // https://datatracker.ietf.org/doc/html/draft-ietf-httpbis-message-signatures-09#section-2.3
-  let signatureBase = ''
-  for (const component of sigInputComponents) {
-    if (component === '@method') {
-      signatureBase += `"@method": ${ctx.request.method}\n`
-    } else if (component === '@target-uri') {
-      signatureBase += `"@target-uri": ${ctx.request.href}\n`
-    } else {
-      signatureBase += `"${component}": ${ctx.headers[component]}\n`
-    }
-  }
-
-  signatureBase += `"@signature-params": ${(
-    ctx.headers['signature-input'] as string
-  )?.replace('sig1=', '')}`
-  return signatureBase
-}
-
-type HttpSigHeaders = Record<'signature' | 'signature-input', string>
-
-type HttpSigRequest = Omit<Context['request'], 'headers'> & {
-  headers: HttpSigHeaders
-}
-
-export type HttpSigContext = Context & {
-  request: HttpSigRequest
-  headers: HttpSigHeaders
-}
-
-function validateHttpSigHeaders(ctx: Context): ctx is HttpSigContext {
-  const sig = ctx.headers['signature']
-  const sigInput = ctx.headers['signature-input'] as string
-
-  const sigInputComponents = getSigInputComponents(sigInput ?? '')
-  if (
-    !sigInputComponents ||
-    !validateSigInputComponents(sigInputComponents, ctx)
-  )
-    return false
-
-  return (
-    sig && sigInput && typeof sig === 'string' && typeof sigInput === 'string'
-  )
 }
 
 export async function grantContinueHttpsigMiddleware(
