+++ packages/http-signature-utils/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "http-message-signatures": "^0.1.2",
     "httpbis-digest-headers": "github:interledger/httpbis-digest-headers",
+    "jose": "^4.9.0",
     "koa": "^2.13.4",
     "koa-bodyparser": "^4.3.0",
     "koa-json": "^2.0.2",
