+++ packages/http-signature-utils/src/index.ts
@@ -1,3 +1,8 @@
 export { generateJwk } from './utils/jwk'
 export { parseOrProvisionKey } from './utils/key'
 export { createSignatureHeaders } from './utils/signatures'
+export {
+  validateHttpSigHeaders,
+  verifySigFromClient,
+  HttpSigContext
+} from './utils/verification'
